# froide/openapi-schema.yaml
openapi: 3.0.3
info:
  title: ''
  version: 0.0.0
paths:
...
  /api/v1/request/{id}/escalate/:
    post:
      operationId: request_escalate_create
      parameters:
      - in: query
        name: format
        schema:
          type: string
          enum:
          - csv
          - json
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this Freedom of Information Request.
        required: true
      tags:
      - request
      requestBody:
        content:
          application/json:
            schema:
              type: object
          application/x-www-form-urlencoded:
            schema:
              type: object
          multipart/form-data:
            schema:
              type: object
        required: true
      security:
      - oauth2:
        - make:request
      - cookieAuth: []
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                type: object
        '400':
          description: ''
...
  /api/v1/request/tags/autocomplete/:
    get:
      operationId: request_tags_autocomplete_retrieve
      parameters:
      - in: query
        name: format
        schema:
          type: string
          enum:
          - csv
          - json
      tags:
      - request
      security:
      - oauth2:
        - make:request
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FoiRequestList'
            text/csv:
              schema:
                $ref: '#/components/schemas/FoiRequestList'
          description: ''
...
components:
  schemas:
    BlankEnum:
      enum:
      - ''
...
    FoiRequestList:
      type: object
      properties:
        resource_uri:
          type: string
          format: uri
          readOnly: true
        id:
          type: integer
          readOnly: true
        url:
          type: string
          readOnly: true
        jurisdiction:
          type: string
          format: uri
          readOnly: true
        is_foi:
          type: boolean
          readOnly: true
          title: Is FoI request
        checked:
          type: boolean
          readOnly: true
        refusal_reason:
          type: string
        costs:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,2})?$
        public:
          type: boolean
          readOnly: true
          title: Published?
        law:
          type: string
          format: uri
        description:
          type: string
        redacted_description:
          type: string
          readOnly: true
        summary:
          type: string
        same_as_count:
          type: integer
          readOnly: true
          title: Identical request count
        same_as:
          type: string
          format: uri
          readOnly: true
        due_date:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        resolved_on:
          type: string
          format: date-time
          readOnly: true
          nullable: true
          title: Resolution date
        last_message:
          type: string
          format: date-time
          readOnly: true
          nullable: true
          title: Date of last message
        created_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        last_modified_at:
          type: string
          format: date-time
          readOnly: true
        status:
          $ref: '#/components/schemas/Status719Enum'
        public_body:
          allOf:
          - $ref: '#/components/schemas/SimplePublicBody'
          readOnly: true
        resolution:
          oneOf:
          - $ref: '#/components/schemas/ResolutionEnum'
          - $ref: '#/components/schemas/BlankEnum'
        slug:
          type: string
          readOnly: true
          pattern: ^[-a-zA-Z0-9_]+$
        title:
          type: string
          readOnly: true
        reference:
          type: string
          readOnly: true
        user:
          type: string
          readOnly: true
        project:
          type: integer
          readOnly: true
        campaign:
          type: string
          format: uri
          readOnly: true
        tags:
          type: array
          items:
            type: string
      required:
      - campaign
      - checked
      - costs
      - created_at
      - description
      - due_date
      - id
      - is_foi
      - jurisdiction
      - last_message
      - last_modified_at
      - law
      - project
      - public
      - public_body
      - redacted_description
      - reference
      - refusal_reason
      - resolved_on
      - resource_uri
      - same_as
      - same_as_count
      - slug
      - status
      - tags
      - title
      - url
      - user
...
    ResolutionEnum:
      enum:
      - successful
      - partially_successful
      - not_held
      - refused
      - user_withdrew_costs
      - user_withdrew
      type: string
      description: |-
        * `successful` - Request Successful
        * `partially_successful` - Request partially successful
        * `not_held` - Information not held
        * `refused` - Request refused
        * `user_withdrew_costs` - Request was withdrawn due to costs
        * `user_withdrew` - Request was withdrawn
...
    SimplePublicBody:
      type: object
      properties:
        resource_uri:
          type: string
          format: uri
          readOnly: true
        id:
          type: integer
        name:
          type: string
          maxLength: 255
        slug:
          type: string
          maxLength: 255
          pattern: ^[-a-zA-Z0-9_]+$
        other_names:
          type: string
        description:
          type: string
        url:
          type: string
          format: uri
          nullable: true
          maxLength: 500
        depth:
          type: integer
          maximum: 32767
          minimum: -32768
        classification:
          type: string
          format: uri
          readOnly: true
        email:
          type: string
          format: email
          maxLength: 255
        contact:
          type: string
        address:
          type: string
        fax:
          type: string
          maxLength: 50
        request_note:
          type: string
        number_of_requests:
          type: integer
          maximum: 2147483647
          minimum: -2147483648
        site_url:
          type: string
        jurisdiction:
          type: string
          format: uri
          readOnly: true
        request_note_html:
          type: string
          readOnly: true
        geo:
          type: string
          readOnly: true
        last_modified_at:
          type: string
      required:
      - classification
      - geo
      - id
      - jurisdiction
      - last_modified_at
      - name
      - request_note_html
      - resource_uri
      - site_url
      - slug
    Status719Enum:
      enum:
      - awaiting_response
      - resolved
      type: string
      description: |-
        * `awaiting_response` - awaiting_response
        * `resolved` - resolved
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
    oauth2:
      type: oauth2
      flows: {}